Server side namespace cleaned
