+++ Assets/Project/Scripts/GameLogic/Server/ServerGameConstants.cs
@@ -1,19 +1,11 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using System;
 
 namespace CEMSIM
 {
-    namespace Network
+    namespace GameLogic
     {
-        namespace Server
+        public class ServerGameConstants
         {
-            public class Constants
-            {
-                // Network Configuration
-                public const int TCP_PORT = 54321;
-                public const int CONCURRENT_CLIENTS = 40;
-
                 // Server Update Rate
                 public const int TICKS_PER_SECOND = 30;
                 public const int MS_PER_TICK = 1000 / TICKS_PER_SECOND; // milliumsecond per tick
@@ -22,8 +14,6 @@ namespace CEMSIM
                 public const float MOVE_SPEED_PER_SECOND = 5f;
                 public const float GRAVITY = -9.81f * 2; // x2 just to make the jump more interesting.
                 public const float JUMP_SPEED_PER_SECOND = 9f;
-
-            }
         }
     }
 }
