+++ Assets/Project/Scripts/GameLogic/Server/ServerGameConstants.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8f573eff4be17492ba7f2a039f62c07e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
