+++ Assets/Project/Scripts/GameLogic/Server/ServerPlayer.cs.meta
