+++ Assets/Project/Scripts/NetworkingScripts/Client/ClientNetworkConstants.cs
@@ -8,7 +8,7 @@ namespace CEMSIM
         {
             public const string SERVER_IP = "orsimserver.ddns.net";
             public const int SERVER_PORT = 54321;
-            public const int DATA_BUFFER_SIZE = 4096; //4MB for both Tx & Rx
+            public const int DATA_BUFFER_SIZE = 4096; //4KB for both Tx & Rx
         }
     }
 }
