+++ Assets/Project/Scripts/NetworkingScripts/ServerMenus/NetworkOverlayMenu.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
-using CEMSIM.Network.Server;
 using CEMSIM.Network;
 using UnityEngine;
 using UnityEngine.UI;
